Binary Tree that appends values from an integer array and it prints out each stem of the binary tree!
ALGORITHM

Takes the number from the array and compares it node values, if node
value is greater then the program sets node pointer to node->bigger and
if the node->bigger is null, then new node is created. If is a smaller
value than the node value then, it checks node->smaller to see if its
available, if its not null, then node is set to node->smaller, and the
function is repeated. Once a null pointer is found that  matches the
condition, it creates a new node at that point.